Automatically popup the method signature helper
We remap '(' in insert mode to:

1. Call omnisharp's signature help.
2. Auto close the parenthesis.
3. Place the cursor in between the parenthesis.

This remapping will also try to give us signature help for if's, for's
while's etc. That can be solved by trying to be clever about what is
contained on the current line but that's a little annoying to do. For
now we just call the signature help method whenever we are in
parenthesis and the only annoying thing about that is that we see the
"No signature help found" message, but it's the truth so no big deal.